requirements.txt
appnope==0.1.2
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap4==4.0.2
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.3
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
traitlets==5.0.5
urllib3==1.26.2
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
